Patch to correctly address spaces in db and query paths
Spaces in db paths are a problem for blast because it makes allows one
to add multiple databases with a double quote. So, to correctly reference
a db path that has spaces, one has to use an outer set of double quotes,
and then an inner set of escaped double quotes. For example,
"\"/path/with a space/to/db\"",
that way if you had two or more DB to
reference, they would be:
"\"/path/with a space/to/db1\" \"/path/with a space/to/db2\"".

More information can be found here: https://www.ncbi.nlm.nih.gov/books/NBK279669/

For the query or subject the same is not applicable. So, one only needs
to make sure it is properly quoted, and shlex.quote does that for us.