Implement click
Remove argparse. This addresses issue #1 suggesting a refactor from argparse
to click. The different workflows
are now called via a flat (-w), and
the FASTA arguments are now at the end of the command
line call, allowing for
patterns like emmtyper *.fasta
to be used.
This also simplified the code by
removing duplication in the
run_ispcr and run_blast
modules from the utilities submodules.